requirements.txt
amqp==5.0.6
APScheduler==3.6.3
asgiref==3.3.4
autopep8==1.5.7
billiard==3.6.4.0
celery==5.0.5
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
cryptography==3.4.7
cycler==0.10.0
Django==3.2
django-filter==2.4.0
django-phonenumber-field==5.1.0
django-telegrambot==1.0.1
djangorestframework==3.12.4
idna==2.10
kiwisolver==1.3.1
kombu==5.0.2
Markdown==3.3.4
matplotlib==3.4.1
mongoengine==0.23.0
numpy==1.20.2
Pillow==8.2.0
prompt-toolkit==3.0.18
psycopg2==2.8.6
pycodestyle==2.7.0
pycparser==2.20
pymongo==3.11.3
pyparsing==2.4.7
python-dateutil==2.8.1
python-decouple==3.4
python-telegram-bot==13.2
pytz==2021.1
requests==2.25.1
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
tornado==6.1
tzlocal==2.1
urllib3==1.26.4
vine==5.0.0
wcwidth==0.2.5
xlwt==1.3.0